Updated Dataset Class, Added Validation, Added Saving
This update changed how ImageDataset provides items from the underlying dataset. Previously, the target data was taken in the shape of the actual image, and then rows and columns were cut in order to match the post-network size of the estimate, so MSE could be taken. Now that trimming is done within ImageDataset, so very little processing needs to be done outside of it. Also added support for all 3 channels to be processed (individually), as well as validation, and saving the model weights after training.